+++ frontend/app/src/electron-main/ipc.ts
@@ -1,16 +1,6 @@
 import { BackendCode } from '@/electron-main/backend-code';
 import { Level } from '@/utils/log-level';
 
-export const IPC_RESTART_BACKEND = 'RESTART_BACKEND' as const;
-export const IPC_CHECK_FOR_UPDATES = 'CHECK_FOR_UPDATES' as const;
-export const IPC_DOWNLOAD_UPDATE = 'DOWNLOAD_UPDATE' as const;
-export const IPC_DOWNLOAD_PROGRESS = 'DOWNLOAD_PROGRESS' as const;
-export const IPC_INSTALL_UPDATE = 'INSTALL_UPDATE' as const;
-export const IPC_DARK_MODE = 'DARK_MODE' as const;
-export const IPC_VERSION = 'VERSION' as const;
-export const IPC_ABOUT = 'ABOUT' as const;
-export const IPC_OPEN_PATH = 'OPEN_PATH' as const;
-
 export type DebugSettings = { vuex: boolean };
 
 type MetamaskImportError = {
@@ -30,6 +20,14 @@ export type SystemVersion = {
   readonly arch: string;
 };
 
+export type BackendOptions = {
+  readonly loglevel: Level;
+  readonly dataDirectory: string;
+  readonly logDirectory: string;
+  readonly sleepSeconds: number;
+  readonly logFromOtherModules: boolean;
+};
+
 export interface Interop {
   openUrl(url: string): Promise<void>;
   openPath(path: string): Promise<void>;
@@ -46,8 +44,9 @@ export interface Interop {
   checkForUpdates(): Promise<boolean>;
   downloadUpdate(progress: (percentage: number) => void): Promise<boolean>;
   installUpdate(): Promise<boolean | Error>;
-  restartBackend(logLevel: Level): Promise<boolean>;
+  restartBackend(options: Partial<BackendOptions>): Promise<boolean>;
   setDarkMode(enabled: boolean): Promise<boolean>;
   version(): Promise<SystemVersion>;
   onAbout(callback: () => void): void;
+  config(defaults: boolean): Promise<Partial<BackendOptions>>;
 }